Add user check to clean_all script.

+++ 99_clean_all.sh
@@ -7,15 +7,23 @@ if [ ! -d "$NATUREDIR" ]; then
     exit 0
 fi
 
-echo -e "\nThis script deletes the working files and directory we created."
+# make sure user wants to delete files 
+read -p "This operation will delete all files in $NATUREDIR. Continue? [n]: " answer 
+if [[ "$answer" == "" ]]; then
+    answer="n"
+fi
 
-cd $NATUREDIR
+if [[ "$answer" == "y" || "$answer" == "Y" ]]; then 
+    cd $NATUREDIR
 
-echo -e "Deleting files."
-rm *.* 
+    echo -e "Deleting files."
+    rm *.txt *.csv *.bak *.cut *.gz 
 
-echo -e "Removing directory $NATUREDIR."
-# move up one directory level in order to delete $NATUREDIR.
-cd ..
-rmdir $NATUREDIR
-echo -e "\nDone!"
+    echo -e "Removing directory $NATUREDIR."
+    # move up one directory level in order to delete $NATUREDIR.
+    cd ..
+    rmdir $NATUREDIR
+    echo -e "\nDone!"
+else 
+    echo -e "No files deleted."
+fi
